OTV: Create new broadcast package
This change creates a broadcast package which contains the broadcast
config. Since this is a base import to many other types and interfaces
this will allow us to start factoring other interfaces out into packages
to clean up the project.
