+++ cmd/oceantv/broadcast.go
@@ -34,10 +34,10 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"strconv"
 	"strings"
 	"time"
 
+	"github.com/ausocean/cloud/cmd/oceantv/broadcast"
 	"github.com/ausocean/cloud/datastore"
 	"github.com/ausocean/cloud/gauth"
 	"github.com/ausocean/cloud/model"
@@ -46,7 +46,7 @@ import (
 type Action int
 
 type (
-	Cfg   = BroadcastConfig
+	Cfg   = broadcast.Config
 	Ctx   = context.Context
 	Store = datastore.Store
 	Key   = datastore.Key
@@ -74,7 +74,6 @@ const (
 
 // Datastore broadcast and live scopes.
 const (
-	broadcastScope            = "Broadcast"                           // Scope under which broadcast configs are stored.
 	liveScope                 = "Live"                                // Scope under which live stream URLs are stored.
 	defaultMessage            = "Welcome to the AusOcean livestream!" // Default message to be sent to the YouTube live chat.
 	tempPin                   = "X60"                                 // Standard temperature pin value.
@@ -85,89 +84,9 @@ const (
 	longTermBroadcastDuration = 1                                     // The duration of the long term broadcast in years.
 )
 
-// BroadcastConfig holds configuration data for a YouTube broadcast.
-type BroadcastConfig struct {
-	UUID                     string        // The immutable unique key of the broadcast.
-	SKey                     int64         // The key of the site this broadcast belongs to.
-	Name                     string        // The name of the broadcast.
-	BID                      string        // Broadcast identification.
-	SID                      string        // Stream ID for any currently associated stream.
-	CID                      string        // ID of associated chat.
-	StreamName               string        // The name of the stream we'll bind to the broadcast.
-	Description              string        // The broadcast description shown below viewing window.
-	LivePrivacy              string        // Privacy of the broadcast whilst live i.e. public, private or unlisted.
-	PostLivePrivacy          string        // Privacy of the broadcast after it has ended i.e. public, private or unlisted.
-	Resolution               string        // Resolution of the stream e.g. 1080p.
-	StartTimestamp           string        // Start time of the broadcast in unix format.
-	Start                    time.Time     // Start time in native go format for easy operations.
-	EndTimestamp             string        // End time of the broadcast in unix format.
-	End                      time.Time     // End time in native go format for easy operations.
-	VidforwardHost           string        // Host address of vidforward service.
-	CameraMac                int64         // Camera hardware's MAC address.
-	ControllerMAC            int64         // Controller hardware's MAC adress (controller used to power camera).
-	OnActions                string        // A series of actions to be used for power up of camera hardware.
-	ShutdownActions          string        // A series of actions to be used for shutdown of camera hardware.
-	OffActions               string        // A series of actions to be used for power down of camera hardware.
-	RTMPVar                  string        // The variable name that holds the RTMP URL and key.
-	Active                   bool          // This is true if the broadcast is currently active i.e. waiting for data or currently streaming.
-	Slate                    bool          // This is true if the broadcast is currently in slate mode i.e. no camera.
-	Issues                   int           // The number of successive stream issues currently experienced. Reset when good health seen.
-	SendMsg                  bool          // True if sensor data will be sent to the YouTube live chat.
-	SensorList               []SensorEntry // List of sensors which can be reported to the YouTube live chat.
-	RTMPKey                  string        // The RTMP key corresponding to the newly created broadcast.
-	UsingVidforward          bool          // Indicates if we're using vidforward i.e. doing long term broadcast.
-	CheckingHealth           bool          // Are we performing health checks for the broadcast? Having this false is useful for dodgy testing streams.
-	AttemptingToStart        bool          // Indicates if we're currently attempting to start the broadcast.
-	Enabled                  bool          // Is the broadcast enabled? If not, it will not be started.
-	Events                   []string      // Holds names of events that are yet to be handled.
-	Unhealthy                bool          // True if the broadcast is unhealthy.
-	BroadcastState           string        // Holds the current state of the broadcast.
-	HardwareState            string        // Holds the current state of the hardware.
-	StartFailures            int           // The number of times the broadcast has failed to start.
-	Transitioning            bool          // If the broadcast is transition from live to slate, or vice versa.
-	StateData                []byte        // States will be marshalled and their data stored here.
-	HardwareStateData        []byte        // Hardware states will be marshalled and their data stored here.
-	Account                  string        // The YouTube account email that this broadcast is associated with.
-	InFailure                bool          // True if the broadcast is in a failure state.
-	BatteryVoltagePin        string        // The pin that the battery voltage is read from.
-	RecoveringVoltage        bool          // True if the broadcast is currently recovering voltage.
-	RequiredStreamingVoltage float64       // The required battery voltage for the camera to stream.
-	VoltageRecoveryTimeout   int           // Max allowable hours for voltage recovery before failure.
-	RegisterOpenFish         bool          // True if the video should be registered with openfish for annotation.
-	OpenFishCaptureSource    string        // The capture source to register the stream to.
-	NotifySuppressRules      string        // Suppression rules for notifications.
-}
-
-func (b *BroadcastConfig) PrettyHardwareStateData() string {
-	return string(b.HardwareStateData)
-}
-
-// SensorEntry contains the information for each sensor.
-type SensorEntry struct {
-	SendMsg   bool
-	Sensor    model.SensorV2
-	Name      string
-	DeviceMac int64
-}
-
 type Camera struct {
 	Name string // Name of camera device.
 	MAC  string // Encoded MAC address of associated camera device.
-}
-
-// parseStartEnd takes the start and end time unix strings from the broadcast
-// and provides these as time.Time.
-func (c *BroadcastConfig) parseStartEnd() error {
-	sInt, err := strconv.ParseInt(c.StartTimestamp, 10, 64)
-	if err != nil {
-		return fmt.Errorf("could not parse unix start time: %w", err)
-	}
-	eInt, err := strconv.ParseInt(c.EndTimestamp, 10, 64)
-	if err != nil {
-		return fmt.Errorf("could not parse unix end time: %w", err)
-	}
-	c.Start, c.End = time.Unix(sInt, 0), time.Unix(eInt, 0)
-	return nil
 }
 
 type oceanTVService struct {
@@ -244,7 +163,7 @@ func (s *oceanTVService) checkBroadcastsHandler(w http.ResponseWriter, r *http.R
 func checkBroadcastsForSites(ctx Ctx, sites []model.Site, eventHooks []eventHook, stateHooks []stateHook) error {
 	var cfgVars []model.Variable
 	for _, s := range sites {
-		vars, err := model.GetVariablesBySite(ctx, store, s.Skey, broadcastScope)
+		vars, err := model.GetVariablesBySite(ctx, store, s.Skey, broadcast.Scope)
 		if err != nil {
 			log.Printf("could not get broadcast entities for site, skey: %d, name: %s, %v", s.Skey, s.Name, err)
 			continue
